Pretty Print Time Duration for Contests (#351)
