+++ vj4/locale/zh_CN.yaml
@@ -187,6 +187,8 @@ This is the current session: 这是当前的会话
 Time Cost: 耗时
 Title: 标题
 Total Time: 总耗时
+Time (Seconds): 耗时 (秒)
+Total Time (Seconds): 总耗时 (秒)
 Type: 类型
 UI Language: 用户界面语言
 Timezone: 时区
