+++ vj4/ui/pages/contest_scoreboard.page.styl
@@ -6,7 +6,7 @@
   .col--user
     width: rem(140px)
 
-  .col--total_score, .col--total_time, .col--solved_problems
+  .col--total_score, .col--total_time_str, .col--solved_problems
     width: rem(100px)
     text-align: center
